fix: nodejs: link package bins after buildPhase
